requirements.txt
discord.py==2.4.0
numpy==1.26.0
python-dotenv==1.0.1
requests==2.32.3
TA-Lib==0.4.31
